+++ home_intent/components/home_assistant/__init__.py
@@ -1,8 +1,8 @@
+from importlib import import_module
 from typing import Set
 
 from pydantic import AnyHttpUrl, BaseModel, Field
 
-from . import cover, fan, group, light, lock, remote, shopping_list, switch, weather
 from .api import HomeAssistantAPI
 
 
@@ -46,39 +46,48 @@ class HomeAssistantComponent:
         self.prefer_toggle = config.prefer_toggle
 
 
-def setup(home_intent):
+def setup(home_intent, language: str):
     # only needed if there are additional settings!
     config = home_intent.get_config(HomeAssistantSettings)
     home_assistant_component = HomeAssistantComponent(config)
 
     if "cover" in home_assistant_component.domains and "cover" not in config.ignore_domains:
+        cover = import_module(f"{__name__}.cover.{language}")
         home_intent.register(cover.Cover(home_assistant_component), cover.intents)
 
     if "fan" in home_assistant_component.domains and "fan" not in config.ignore_domains:
+        fan = import_module(f"{__name__}.fan.{language}")
         home_intent.register(fan.Fan(home_assistant_component), fan.intents)
 
     if "group" not in config.ignore_domains:
+        group = import_module(f"{__name__}.group.{language}")
         home_intent.register(group.Group(home_assistant_component), group.intents)
 
     if "light" in home_assistant_component.domains and "light" not in config.ignore_domains:
+        light = import_module(f"{__name__}.light.{language}")
         home_intent.register(light.Light(home_assistant_component), light.intents)
 
     if "lock" in home_assistant_component.domains and "lock" not in config.ignore_domains:
+        lock = import_module(f"{__name__}.lock.{language}")
         home_intent.register(lock.Lock(home_assistant_component), lock.intents)
 
     if "remote" in home_assistant_component.domains and "remote" not in config.ignore_domains:
+        remote = import_module(f"{__name__}.remote.{language}")
         home_intent.register(remote.Remote(home_assistant_component), remote.intents)
 
     if (
         "shopping_list" in home_assistant_component.domains
         and "shopping_list" not in config.ignore_domains
     ):
+        shopping_list = import_module(f"{__name__}.shopping_list.{language}")
         home_intent.register(
             shopping_list.ShoppingList(home_assistant_component), shopping_list.intents
         )
 
     if "switch" in home_assistant_component.domains and "switch" not in config.ignore_domains:
+        switch = import_module(f"{__name__}.switch.{language}")
         home_intent.register(switch.Switch(home_assistant_component), switch.intents)
 
     if "weather" in home_assistant_component.domains and "weather" not in config.ignore_domains:
+        weather = import_module(f"{__name__}.weather.{language}")
         home_intent.register(weather.Weather(home_assistant_component), weather.intents)
