+++ home_intent/components/home_assistant/cover/base_cover.py
@@ -16,7 +16,7 @@ class SupportedFeatures(IntFlag):
     SUPPORT_STOP_TILT = auto()
 
 
-class Cover:
+class BaseCover:
     def __init__(self, home_assistant):
         self.ha = home_assistant
         self.entities = [x for x in self.ha.entities if x["entity_id"].startswith("cover.")]
@@ -97,52 +97,41 @@ class Cover:
         slots = {"half way": 50}
         return slots
 
-    @intents.sentences(["open [the] ($cover_open)"])
-    def open_cover(self, cover_open):
+    def _open_cover(self, cover_open):
         if SupportedFeatures.SUPPORT_OPEN in self.cover_features[cover_open]:
             self.ha.api.call_service("cover", "open_cover", {"entity_id": cover_open})
         elif SupportedFeatures.SUPPORT_OPEN_TILT in self.cover_features[cover_open]:
             self.ha.api.call_service("cover", "open_cover_tilt", {"entity_id": cover_open})
         response = self.ha.api.get_entity(cover_open)
-        return f"Opening the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["close [the] ($cover_close)"])
-    def close_cover(self, cover_close):
+    def _close_cover(self, cover_close):
         if SupportedFeatures.SUPPORT_CLOSE in self.cover_features[cover_close]:
             self.ha.api.call_service("cover", "close_cover", {"entity_id": cover_close})
         elif SupportedFeatures.SUPPORT_CLOSE_TILT in self.cover_features[cover_close]:
             self.ha.api.call_service("cover", "close_cover_tilt", {"entity_id": cover_close})
         response = self.ha.api.get_entity(cover_close)
-        return f"Closing the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["stop [the] ($cover_stop)"])
-    def stop_cover(self, cover_stop):
+    def _stop_cover(self, cover_stop):
         if SupportedFeatures.SUPPORT_STOP in self.cover_features[cover_stop]:
             self.ha.api.call_service("cover", "stop_cover", {"entity_id": cover_stop})
         if SupportedFeatures.SUPPORT_STOP_TILT in self.cover_features[cover_stop]:
             self.ha.api.call_service("cover", "stop_cover_tilt", {"entity_id": cover_stop})
         response = self.ha.api.get_entity(cover_stop)
-        return f"Stopping the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["tilt open [the] ($cover_open_tilt)"])
-    def open_cover_tilt(self, cover_open_tilt):
+    def _open_cover_tilt(self, cover_open_tilt):
         self.ha.api.call_service("cover", "open_cover_tilt", {"entity_id": cover_open_tilt})
         response = self.ha.api.get_entity(cover_open_tilt)
-        return f"Opening the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["tilt close [the] ($cover_close_tilt)"])
-    def close_cover_tilt(self, cover_close_tilt):
+    def _close_cover_tilt(self, cover_close_tilt):
         self.ha.api.call_service("cover", "close_cover_tilt", {"entity_id": cover_close_tilt})
         response = self.ha.api.get_entity(cover_close_tilt)
-        return f"Closing the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(
-        [
-            "(set|change|make) [the] ($cover_set_position) [position] [to] ($cover_positions:!int) [percent]",
-            "(open|close) [the] ($cover_set_position) [to] ($cover_positions:!int) [percent]",
-        ]
-    )
-    def set_cover_position(self, cover_set_position, cover_positions):
+    def _set_cover_position(self, cover_set_position, cover_positions):
         if SupportedFeatures.SUPPORT_SET_POSITION:
             self.ha.api.call_service(
                 "cover",
@@ -156,19 +145,13 @@ class Cover:
                 {"entity_id": cover_set_position, "position": cover_positions},
             )
         response = self.ha.api.get_entity(cover_set_position)
-        return f"Setting the {response['attributes']['friendly_name']} to {cover_positions}%"
+        return response
 
-    @intents.sentences(
-        [
-            "(set|change|make) [the] ($cover_set_tilt_position) tilt [position] [to] ($cover_positions:!int) [percent]",
-            "tilt (open|close) [the] ($cover_set_tilt_position) [to] ($cover_positions:!int) [percent]",
-        ]
-    )
-    def set_cover_tilt_position(self, cover_set_tilt_position, cover_positions):
+    def _set_cover_tilt_position(self, cover_set_tilt_position, cover_positions):
         self.ha.api.call_service(
             "cover",
             "set_cover_tilt_position",
             {"entity_id": cover_set_tilt_position, "tilt_position": cover_positions},
         )
         response = self.ha.api.get_entity(cover_set_tilt_position)
-        return f"Setting the {response['attributes']['friendly_name']} to {cover_positions}%"
+        return response
