+++ home_intent/components/home_assistant/group/base_group.py
@@ -3,7 +3,7 @@ from home_intent import Intents
 intents = Intents(__name__)
 
 
-class Group:
+class BaseGroup:
     def __init__(self, home_assistant):
         self.ha = home_assistant
         self.entities = [x for x in self.ha.entities if x["entity_id"].startswith("group.")]
@@ -17,20 +17,17 @@ class Group:
         }
         return slots
 
-    @intents.sentences(["toggle [the] ($group)"])
-    def toggle_group(self, group):
+    def _toggle_group(self, group):
         self.ha.api.call_service("homeassistant", "toggle", {"entity_id": group})
         response = self.ha.api.get_entity(group)
-        return f"Toggling the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["turn on [the] ($group)"])
-    def turn_on_group(self, group):
+    def _turn_on_group(self, group):
         self.ha.api.call_service("homeassistant", "turn_on", {"entity_id": group})
         response = self.ha.api.get_entity(group)
-        return f"Turning on the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["turn off [the] ($group)"])
-    def turn_off_group(self, group):
+    def _turn_off_group(self, group):
         self.ha.api.call_service("homeassistant", "turn_off", {"entity_id": group})
         response = self.ha.api.get_entity(group)
-        return f"Turning off the {response['attributes']['friendly_name']}"
+        return response
