+++ home_intent/components/home_assistant/shopping_list/base_shopping_list.py
