+++ home_intent/components/home_assistant/switch/base_switch.py
@@ -3,7 +3,7 @@ from home_intent import Intents
 intents = Intents(__name__)
 
 
-class Switch:
+class BaseSwitch:
     def __init__(self, home_assistant):
         self.ha = home_assistant
         self.entities = [x for x in self.ha.entities if x["entity_id"].startswith("switch.")]
@@ -17,28 +17,17 @@ class Switch:
         }
         return slots
 
-    @intents.on_event("register_sentences")
-    def handle_prefer_toggle(self):
-        if self.ha.prefer_toggle:
-            intents.disable_intent(self.turn_on)
-            intents.disable_intent(self.turn_off)
-        else:
-            intents.disable_intent(self.toggle_switch)
-
-    @intents.sentences(["toggle the ($switch)", "turn (on | off) [the] ($switch)"])
-    def toggle_switch(self, switch):
+    def _toggle_switch(self, switch):
         self.ha.api.call_service("switch", "toggle", {"entity_id": switch})
         response = self.ha.api.get_entity(switch)
-        return f"Turning {response['state']} the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["turn off [the] ($switch)"])
-    def turn_off(self, switch):
+    def _turn_off(self, switch):
         self.ha.api.call_service("switch", "turn_off", {"entity_id": switch})
         response = self.ha.api.get_entity(switch)
-        return f"Turning {response['state']} the {response['attributes']['friendly_name']}"
+        return response
 
-    @intents.sentences(["turn on [the] ($switch)"])
-    def turn_on(self, switch):
+    def _turn_on(self, switch):
         self.ha.api.call_service("switch", "turn_on", {"entity_id": switch})
         response = self.ha.api.get_entity(switch)
-        return f"Turning {response['state']} the {response['attributes']['friendly_name']}"
+        return response
