+++ home_intent/components/home_assistant/weather/base_weather.py
