+++ home_intent/components/timer/__init__.py
@@ -1,86 +1,7 @@
-from collections import defaultdict
-from datetime import timedelta
-from threading import Timer as ThreadingTimer
-
-from humanize import precisedelta
-from pydantic import BaseModel
-
+from importlib import import_module
 from home_intent import HomeIntent, Intents
 
-intents = Intents(__name__)
 
-
-class TimerException(Exception):
-    pass
-
-
-class Timer:
-    def __init__(self, home_intent: HomeIntent):
-        # TODO: keep track of timers and add ability to remove timers
-        # self.timers = []
-        self.home_intent = home_intent
-
-    @intents.dictionary_slots
-    def partial_time(self):
-        return {
-            "and [a] half": "half",
-            "and [a] quarter": "quarter",
-            "and [a] third": "third",
-        }
-
-    @intents.sentences(
-        [
-            "time = 0..128",
-            "set timer [(<time>){hours} hours] [(<time>){minutes} minutes] [(<time>){seconds} seconds]",
-            "set timer (<time>){hours} [($partial_time)] hours",
-            "set timer (<time>){minutes} [($partial_time)] minutes",
-            "set timer (<time>){seconds} [($partial_time)] seconds",
-            "set a [(<time>){hours} hour] [(<time>){minutes} minute] [(<time>){seconds} second] timer",
-            "set a (<time>){hours} [($partial_time)] hour timer",
-            "set a (<time>){minutes} [($partial_time)] minute timer",
-            "set a (<time>){seconds} [($partial_time)] second timer",
-        ]
-    )
-    def set_timer(
-        self, hours: int = None, minutes: int = None, seconds: int = None, partial_time=None
-    ):
-        timer_duration = timedelta(
-            hours=int(hours or 0), minutes=int(minutes or 0), seconds=int(seconds or 0),
-        )
-        if timer_duration == timedelta(0):
-            raise TimerException("Timer has to be set for more than 0 seconds")
-        if partial_time:
-            timer_duration = timer_duration + get_partial_time_duration(
-                partial_time, hours, minutes, seconds
-            )
-        human_timer_duration = precisedelta(timer_duration)
-        timer = ThreadingTimer(
-            timer_duration.total_seconds(), self.complete_timer, (human_timer_duration,),
-        )
-        timer.start()
-        return f"Setting timer {human_timer_duration}"
-
-    def complete_timer(self, human_timer_duration: str):
-        self.home_intent.play_audio_file("timer/alarm.wav")
-        self.home_intent.say(f"Your timer {human_timer_duration} has ended")
-
-
-def get_partial_time_duration(partial_time, hours=None, minutes=None, seconds=None):
-    partial_of = None
-    if hours:
-        partial_of = "hours"
-    elif minutes:
-        partial_of = "minutes"
-    elif seconds:
-        partial_of = "seconds"
-
-    if partial_time == "half":
-        return timedelta(**{partial_of: 0.5})
-    elif partial_time == "quarter":
-        return timedelta(**{partial_of: 0.25})
-    elif partial_time == "third":
-        return timedelta(**{partial_of: 1 / 3})
-
-
-def setup(home_intent: HomeIntent):
-    home_intent.register(Timer(home_intent), intents)
+def setup(home_intent: HomeIntent, language: str):
+    timer = import_module(f"{__name__}.{language}")
+    home_intent.register(timer.Timer(home_intent, language), timer.intents)
